requirements.txt
click~=8.0.1
telethon~=1.22.0
filesplit~=3.0.2
pysocks~=1.7.1